[#312] refactor: 로그아웃 버그, 검색 인풋 초기화, 고객센터 제보 버튼 추가
