+++ src/shared/ui/scroll-top-button.tsx
