+++ src/views/search/ui/search-page.tsx
@@ -1,6 +1,6 @@
 import SearchInput from '@/features/search/ui/search-input';
 import SearchResults from '@/widgets/search/ui/search-results';
-import ScrollTopButton from '@/features/recruit/ui/scroll-top-button';
+import ScrollTopButton from '@/shared/ui/scroll-top-button';
 import { Suspense } from 'react';
 import SearchListSkeletonLoading from '@/entities/search/ui/search-list-skeleton-loading';
