+++ src/views/support/ui/support-page.tsx
@@ -1,13 +1,13 @@
-import ScrollTopButton from '@/features/recruit/ui/scroll-top-button';
 import Search from '@/features/support/ui/search';
 import FAQList from '@/widgets/support/faq-list';
+import BugInfoButton from '@/features/support/ui/bug-info-button';
 
 function SupportPage() {
   return (
     <>
       <Search />
       <FAQList />
-      <ScrollTopButton />
+      <BugInfoButton />
     </>
   );
 }
