+++ src/widgets/support/faq-list.tsx
@@ -85,7 +85,7 @@ function FAQList() {
         <div key={item.question} className="border-b pb-4">
           <button
             className={cn(
-              'w-full text-left text-lg font-semibold transition hover:text-[#20E86C]',
+              'w-full cursor-pointer text-left text-lg font-semibold transition hover:text-[#20E86C]',
               openIndex === idx ? 'text-[#20E86C]' : 'text-black',
             )}
             onClick={() => toggle(idx)}
@@ -109,6 +109,7 @@ function FAQList() {
 
       <span className="text-xl font-bold text-[#20E86C] underline">
         <a
+          id="bottom"
           href="https://docs.google.com/forms/d/1U3tnUQxdu8CtT9W1b8KQoQkhFs851BiwSO3l355L57w/viewform?edit_requested=true"
           target="_blank"
           rel="noopener noreferrer"
